Move definition of `improve_hash_formatting`.
We don’t want it defined in the `Pretty` module because
that is included the custom matcher DSL and it’s not meant
for end-user use.
